fix: fixed some endpoints feat: added comments to every endpoint that didn't have them fix: fixed authory of some endpoints

+++ src/http/controllers/users/delete.ts
@@ -5,6 +5,10 @@ import { User } from "../../../models/user";
 import authMiddleware from "../../middlewares/auth";
 import adminMiddleware from "../../middlewares/admin";
 
+// Autora: Lucía Lozano López
+//
+// Esta ruta recibirá un ID en parámetros y buscará el usuario para eliminarlo.
+// Hace uso de los middlewares de autenticación y administración.
 export default (app: OpenAPIHono) => {
   app.use("/users/:id", authMiddleware);
   app.use("/users/:id", adminMiddleware);
@@ -35,10 +39,23 @@ export default (app: OpenAPIHono) => {
     }),
 
     async function (c: Context): Promise<any> {
+      const id = c.req.param("id");
+
+      // Se recoge el ID de la URL y se comprueba que exista en los parámetros de la URL
+      if (!id) {
+        return c.json({
+          message: "No params provided"
+        },
+          400
+        );
+      }
+
+      // Se busca por su ID y se elimina, devolviendo el usuario eliminado
       const user = await User.findByIdAndDelete({
-        _id: c.req.param("id"),
+        id,
       });
 
+      // Se comprueba que se haya eliminado
       if (!user) {
         return c.json(
           {
@@ -47,11 +64,13 @@ export default (app: OpenAPIHono) => {
           404
         );
       }
+
+      // Se devuelve un código 204 para indicar éxito
       return c.json(
         {
           message: "User deleted",
         },
-        200
+        204
       );
     }
   );
